Backend/requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
email-validator==2.1.0

flask-mail==0.9.1       # For email notifications
gunicorn==21.2.0        # For production deployment
pytest==7.4.3           # For testing